# marketplace/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: Marketplace Service API
  description: |
    API for managing the marketplace ecosystem including vendors, product listings, reviews, and marketplace analytics.
    This service handles vendor management, product listings, reviews, and marketplace operations.
  version: 1.0.0
  contact:
    name: Marketplace Service Support
    email: support@microsocialecommerce.com
servers:
  - url: http://localhost:8083
    description: Local development server
  - url: https://api.microsocialecommerce.com/marketplace
    description: Production server
...
paths:
...
  /api/vendors/{id}/approve:
    post:
      tags:
        - Vendor Lifecycle
      summary: Approve vendor
      description: Approve a vendor application (Admin/Marketplace Manager only)
      operationId: approveVendor
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Vendor ID
          schema:
            type: integer
            format: int64
        - name: approvedBy
          in: query
          required: true
          description: Administrator approving the vendor
          schema:
            type: string
      responses:
        '200':
          description: Vendor approved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Vendor'
        '403':
          description: Access denied
        '404':
          description: Vendor not found
  /api/vendors/{id}/reject:
    post:
      tags:
        - Vendor Lifecycle
      summary: Reject vendor
      description: Reject a vendor application (Admin/Marketplace Manager only)
      operationId: rejectVendor
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Vendor ID
          schema:
            type: integer
            format: int64
        - name: reason
          in: query
          required: true
          description: Reason for rejection
          schema:
            type: string
      responses:
        '200':
          description: Vendor rejected successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Vendor'
        '403':
          description: Access denied
        '404':
          description: Vendor not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
...
    Vendor:
      type: object
      properties:
        id:
          type: integer
          format: int64
        seller:
          $ref: '#/components/schemas/Seller'
        companyName:
          type: string
        legalName:
          type: string
        registrationNumber:
          type: string
        taxId:
          type: string
        website:
          type: string
        contactName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        businessAddress:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        postalCode:
          type: string
        yearEstablished:
          type: integer
        numberOfEmployees:
          type: integer
        annualRevenue:
          type: string
        vendorType:
          $ref: '#/components/schemas/VendorType'
        status:
          $ref: '#/components/schemas/VendorStatus'
        notes:
          type: string
        onboardingCompleted:
          type: boolean
        documentsVerified:
          type: boolean
        termsAccepted:
          type: boolean
        termsAcceptedDate:
          type: string
          format: date
        agreementSignedDate:
          type: string
          format: date
        kycVerified:
          type: boolean
        approvalDate:
          type: string
          format: date
        approvedBy:
          type: string
        acceptsReturns:
          type: boolean
        returnPolicy:
          type: string
        offersInternationalShipping:
          type: boolean
        hasTaxExemption:
          type: boolean
        supportEmail:
          type: string
        supportPhone:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Seller:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
...
    VendorStatus:
      type: string
      enum:
        - PENDING_APPROVAL
        - ACTIVE
        - INACTIVE
        - SUSPENDED
        - TERMINATED
        - REJECTED
    VendorType:
      type: string
      enum:
        - INDIVIDUAL
        - SMALL_BUSINESS
        - MEDIUM_BUSINESS
        - ENTERPRISE
        - DISTRIBUTOR
        - WHOLESALER
        - MANUFACTURER
...