# vendor-onboarding/src/main/java/com/exalt/ecosystem/socialcommerce/vendoronboarding/controller/VendorOnboardingController.java
package com.exalt.ecosystem.socialcommerce.vendoronboarding.controller;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;

import java.time.LocalDateTime;
import java.util.*;

/**
 * Vendor Onboarding Controller
 * 
 * Manages vendor registration, KYC verification, document upload, and onboarding workflows
 * for the Social Commerce Ecosystem.
 */
@RestController
@RequestMapping("/api/v1/vendor-onboarding")
@Tag(name = "Vendor Onboarding", description = "APIs for vendor registration and KYC verification")
public class VendorOnboardingController {

    @Operation(summary = "Start vendor registration", description = "Initiates the vendor onboarding process with basic business information")
    @PostMapping("/register")
    public ResponseEntity<Map<String, Object>> startVendorRegistration(@RequestBody Map<String, Object> registrationRequest) {
        try {
            String businessName = (String) registrationRequest.get("businessName");
            String contactEmail = (String) registrationRequest.get("contactEmail");
            String businessType = (String) registrationRequest.get("businessType");
            String registrationNumber = (String) registrationRequest.get("registrationNumber");
            
            Map<String, Object> vendor = new HashMap<>();
            vendor.put("vendorId", "VND_" + System.currentTimeMillis());
            vendor.put("businessName", businessName);
            vendor.put("contactEmail", contactEmail);
            vendor.put("businessType", businessType);
            vendor.put("registrationNumber", registrationNumber);
            vendor.put("onboardingStatus", "REGISTRATION_STARTED");
            vendor.put("registrationDate", LocalDateTime.now());
            vendor.put("kycStatus", "PENDING");
            vendor.put("complianceStatus", "PENDING");
            vendor.put("applicationReference", "APP_" + System.currentTimeMillis());
            
            Map<String, Object> response = new HashMap<>();
            response.put("success", true);
            response.put("message", "Vendor registration started successfully");
            response.put("vendor", vendor);
            response.put("nextSteps", Arrays.asList(
                "Complete business profile",
                "Upload required documents",
                "Submit KYC information",
                "Await verification"
            ));
            response.put("timestamp", LocalDateTime.now());
            
            return ResponseEntity.ok(response);
        } catch (Exception e) {
            Map<String, Object> errorResponse = new HashMap<>();
            errorResponse.put("success", false);
            errorResponse.put("error", "Failed to start registration: " + e.getMessage());
            errorResponse.put("timestamp", LocalDateTime.now());
            
            return ResponseEntity.badRequest().body(errorResponse);
        }
    }

    @Operation(summary = "Complete business profile", description = "Updates vendor business profile with detailed information")
    @PutMapping("/{vendorId}/profile")
    public ResponseEntity<Map<String, Object>> completeBusinessProfile(
            @PathVariable String vendorId,
            @RequestBody Map<String, Object> profileRequest) {
        
        Map<String, Object> businessProfile = new HashMap<>();
        businessProfile.put("vendorId", vendorId);
        businessProfile.put("businessDescription", profileRequest.get("businessDescription"));
        businessProfile.put("businessAddress", profileRequest.get("businessAddress"));
        businessProfile.put("businessPhone", profileRequest.get("businessPhone"));
        businessProfile.put("businessWebsite", profileRequest.get("businessWebsite"));
        businessProfile.put("yearEstablished", profileRequest.get("yearEstablished"));
        businessProfile.put("employeeCount", profileRequest.get("employeeCount"));
        businessProfile.put("annualRevenue", profileRequest.get("annualRevenue"));
        businessProfile.put("productCategories", profileRequest.get("productCategories"));
        businessProfile.put("shippingLocations", profileRequest.get("shippingLocations"));
        businessProfile.put("onboardingStatus", "PROFILE_COMPLETED");
        businessProfile.put("updatedAt", LocalDateTime.now());
        
        Map<String, Object> response = new HashMap<>();
        response.put("success", true);
        response.put("message", "Business profile completed successfully");
        response.put("businessProfile", businessProfile);
        response.put("nextStep", "Upload required documents");
        response.put("timestamp", LocalDateTime.now());
        
        return ResponseEntity.ok(response);
    }

    @Operation(summary = "Upload verification documents", description = "Uploads business documents required for KYC verification")
    @PostMapping("/{vendorId}/documents")
    public ResponseEntity<Map<String, Object>> uploadDocuments(
            @PathVariable String vendorId,
            @RequestParam("documentType") String documentType,
            @RequestParam("file") MultipartFile file) {
        
        try {
            // Simulate document processing
            Map<String, Object> documentInfo = new HashMap<>();
            documentInfo.put("documentId", "DOC_" + System.currentTimeMillis());
            documentInfo.put("vendorId", vendorId);
            documentInfo.put("documentType", documentType);
            documentInfo.put("fileName", file.getOriginalFilename());
            documentInfo.put("fileSize", file.getSize());
            documentInfo.put("uploadedAt", LocalDateTime.now());
            documentInfo.put("verificationStatus", "UNDER_REVIEW");
            documentInfo.put("downloadUrl", "/api/v1/vendor-onboarding/documents/" + documentInfo.get("documentId"));
            
            Map<String, Object> response = new HashMap<>();
            response.put("success", true);
            response.put("message", "Document uploaded successfully");
            response.put("document", documentInfo);
            response.put("estimatedReviewTime", "2-3 business days");
            response.put("timestamp", LocalDateTime.now());
            
            return ResponseEntity.ok(response);
        } catch (Exception e) {
            Map<String, Object> errorResponse = new HashMap<>();
            errorResponse.put("success", false);
            errorResponse.put("error", "Failed to upload document: " + e.getMessage());
            errorResponse.put("timestamp", LocalDateTime.now());
            
            return ResponseEntity.badRequest().body(errorResponse);
        }
    }

    @Operation(summary = "Submit KYC information", description = "Submits Know Your Customer information for business verification")
    @PostMapping("/{vendorId}/kyc")
    public ResponseEntity<Map<String, Object>> submitKYC(
            @PathVariable String vendorId,
            @RequestBody Map<String, Object> kycRequest) {
        
        Map<String, Object> kycInfo = new HashMap<>();
        kycInfo.put("vendorId", vendorId);
        kycInfo.put("businessOwnerName", kycRequest.get("businessOwnerName"));
        kycInfo.put("ownerId", kycRequest.get("ownerId"));
        kycInfo.put("ownerNationality", kycRequest.get("ownerNationality"));
        kycInfo.put("businessLicense", kycRequest.get("businessLicense"));
        kycInfo.put("taxId", kycRequest.get("taxId"));
        kycInfo.put("bankAccountDetails", kycRequest.get("bankAccountDetails"));
        kycInfo.put("authorizedSignatory", kycRequest.get("authorizedSignatory"));
        kycInfo.put("kycStatus", "SUBMITTED");
        kycInfo.put("submittedAt", LocalDateTime.now());
        kycInfo.put("verificationLevel", "LEVEL_2");
        
        Map<String, Object> response = new HashMap<>();
        response.put("success", true);
        response.put("message", "KYC information submitted successfully");
        response.put("kycInfo", kycInfo);
        response.put("verificationSteps", Arrays.asList(
            "Document authenticity check",
            "Business registration verification",
            "Bank account verification",
            "Identity verification"
        ));
        response.put("timestamp", LocalDateTime.now());
        
        return ResponseEntity.ok(response);
    }

    @Operation(summary = "Get onboarding status", description = "Retrieves current onboarding status and progress for a vendor")
    @GetMapping("/{vendorId}/status")
    public ResponseEntity<Map<String, Object>> getOnboardingStatus(@PathVariable String vendorId) {
        Map<String, Object> onboardingStatus = new HashMap<>();
        onboardingStatus.put("vendorId", vendorId);
        onboardingStatus.put("overallStatus", "IN_PROGRESS");
        onboardingStatus.put("progressPercentage", 75);
        
        Map<String, Object> stepStatus = new HashMap<>();
        stepStatus.put("registration", "COMPLETED");
        stepStatus.put("businessProfile", "COMPLETED");
        stepStatus.put("documentUpload", "COMPLETED");
        stepStatus.put("kycVerification", "IN_PROGRESS");
        stepStatus.put("complianceCheck", "PENDING");
        stepStatus.put("finalApproval", "PENDING");
        
        onboardingStatus.put("stepStatus", stepStatus);
        onboardingStatus.put("estimatedCompletionTime", "3-5 business days");
        onboardingStatus.put("lastUpdated", LocalDateTime.now());
        
        Map<String, Object> response = new HashMap<>();
        response.put("success", true);
        response.put("onboardingStatus", onboardingStatus);
        response.put("timestamp", LocalDateTime.now());
        
        return ResponseEntity.ok(response);
    }

    @Operation(summary = "Approve vendor application", description = "Approves vendor application after successful verification (Admin only)")
    @PostMapping("/{vendorId}/approve")
    public ResponseEntity<Map<String, Object>> approveVendor(
            @PathVariable String vendorId,
            @RequestBody Map<String, Object> approvalRequest) {
        
        Map<String, Object> approval = new HashMap<>();
        approval.put("vendorId", vendorId);
        approval.put("approvedBy", approvalRequest.get("approvedBy"));
        approval.put("approvalDate", LocalDateTime.now());
        approval.put("approvalNotes", approvalRequest.get("approvalNotes"));
        approval.put("onboardingStatus", "APPROVED");
        approval.put("marketplaceStatus", "ACTIVE");
        approval.put("vendorTier", "STANDARD");
        approval.put("commissionRate", "5.0%");
        approval.put("payoutSchedule", "WEEKLY");
        
        Map<String, Object> response = new HashMap<>();
        response.put("success", true);
        response.put("message", "Vendor application approved successfully");
        response.put("approval", approval);
        response.put("nextSteps", Arrays.asList(
            "Marketplace account activated",
            "Product listing access granted",
            "Payment setup completed",
            "Welcome email sent"
        ));
        response.put("timestamp", LocalDateTime.now());
        
        return ResponseEntity.ok(response);
    }

    @Operation(summary = "Reject vendor application", description = "Rejects vendor application with reason (Admin only)")
    @PostMapping("/{vendorId}/reject")
    public ResponseEntity<Map<String, Object>> rejectVendor(
            @PathVariable String vendorId,
            @RequestBody Map<String, Object> rejectionRequest) {
        
        Map<String, Object> rejection = new HashMap<>();
        rejection.put("vendorId", vendorId);
        rejection.put("rejectedBy", rejectionRequest.get("rejectedBy"));
        rejection.put("rejectionDate", LocalDateTime.now());
        rejection.put("rejectionReason", rejectionRequest.get("rejectionReason"));
        rejection.put("onboardingStatus", "REJECTED");
        rejection.put("canReapply", true);
        rejection.put("reapplyAfter", LocalDateTime.now().plusDays(30));
        
        Map<String, Object> response = new HashMap<>();
        response.put("success", true);
        response.put("message", "Vendor application rejected");
        response.put("rejection", rejection);
        response.put("supportMessage", "You can reapply after addressing the mentioned issues");
        response.put("timestamp", LocalDateTime.now());
        
        return ResponseEntity.ok(response);
    }

    @Operation(summary = "Get pending applications", description = "Retrieves list of vendor applications pending review (Admin only)")
    @GetMapping("/pending")
    public ResponseEntity<Map<String, Object>> getPendingApplications(
            @RequestParam(defaultValue = "0") int page,
            @RequestParam(defaultValue = "10") int size) {
        
        List<Map<String, Object>> pendingApplications = Arrays.asList(
            createSampleApplication("VND_001", "TechCorp Solutions", "IN_PROGRESS", 75),
            createSampleApplication("VND_002", "Fashion Forward Ltd", "KYC_REVIEW", 60),
            createSampleApplication("VND_003", "Global Electronics", "DOCUMENT_REVIEW", 45),
            createSampleApplication("VND_004", "Organic Foods Co", "COMPLIANCE_CHECK", 80),
            createSampleApplication("VND_005", "Sports Gear Inc", "FINAL_REVIEW", 90)
        );
        
        Map<String, Object> response = new HashMap<>();
        response.put("success", true);
        response.put("pendingApplications", pendingApplications);
        response.put("totalCount", pendingApplications.size());
        response.put("page", page);
        response.put("size", size);
        response.put("averageProcessingTime", "4.2 days");
        response.put("timestamp", LocalDateTime.now());
        
        return ResponseEntity.ok(response);
    }

    @Operation(summary = "Get onboarding analytics", description = "Provides analytics for vendor onboarding performance")
    @GetMapping("/analytics")
    public ResponseEntity<Map<String, Object>> getOnboardingAnalytics() {
        Map<String, Object> analytics = new HashMap<>();
        analytics.put("totalApplications", 1245);
        analytics.put("approvedVendors", 987);
        analytics.put("rejectedApplications", 158);
        analytics.put("pendingApplications", 100);
        analytics.put("averageProcessingTime", "4.2 days");
        analytics.put("approvalRate", "79.2%");
        analytics.put("applicationVolume30Days", 156);
        
        Map<String, Object> statusDistribution = new HashMap<>();
        statusDistribution.put("REGISTRATION_STARTED", 25);
        statusDistribution.put("PROFILE_COMPLETED", 20);
        statusDistribution.put("DOCUMENT_UPLOAD", 15);
        statusDistribution.put("KYC_REVIEW", 18);
        statusDistribution.put("COMPLIANCE_CHECK", 12);
        statusDistribution.put("FINAL_REVIEW", 10);
        
        analytics.put("statusDistribution", statusDistribution);
        
        Map<String, Object> rejectionReasons = new HashMap<>();
        rejectionReasons.put("INCOMPLETE_DOCUMENTS", 35);
        rejectionReasons.put("FAILED_KYC", 28);
        rejectionReasons.put("COMPLIANCE_ISSUES", 22);
        rejectionReasons.put("BUSINESS_MODEL_MISMATCH", 15);
        
        analytics.put("rejectionReasons", rejectionReasons);
        
        Map<String, Object> response = new HashMap<>();
        response.put("success", true);
        response.put("analytics", analytics);
        response.put("timestamp", LocalDateTime.now());
        
        return ResponseEntity.ok(response);
    }

    @Operation(summary = "Update vendor tier", description = "Updates vendor tier based on performance (Admin only)")
    @PutMapping("/{vendorId}/tier")
    public ResponseEntity<Map<String, Object>> updateVendorTier(
            @PathVariable String vendorId,
            @RequestBody Map<String, Object> tierRequest) {
        
        Map<String, Object> tierUpdate = new HashMap<>();
        tierUpdate.put("vendorId", vendorId);
        tierUpdate.put("previousTier", "STANDARD");
        tierUpdate.put("newTier", tierRequest.get("newTier"));
        tierUpdate.put("updatedBy", tierRequest.get("updatedBy"));
        tierUpdate.put("updateReason", tierRequest.get("updateReason"));
        tierUpdate.put("updatedAt", LocalDateTime.now());
        tierUpdate.put("newCommissionRate", calculateCommissionRate((String) tierRequest.get("newTier")));
        tierUpdate.put("newBenefits", getTierBenefits((String) tierRequest.get("newTier")));
        
        Map<String, Object> response = new HashMap<>();
        response.put("success", true);
        response.put("message", "Vendor tier updated successfully");
        response.put("tierUpdate", tierUpdate);
        response.put("timestamp", LocalDateTime.now());
        
        return ResponseEntity.ok(response);
    }

    /**
     * Helper method to create sample application object
     */
    private Map<String, Object> createSampleApplication(String vendorId, String businessName, String status, int progress) {
        Map<String, Object> application = new HashMap<>();
        application.put("vendorId", vendorId);
        application.put("businessName", businessName);
        application.put("onboardingStatus", status);
        application.put("progressPercentage", progress);
        application.put("submittedAt", LocalDateTime.now().minusDays((long) (Math.random() * 30)));
        application.put("businessType", "LIMITED_COMPANY");
        application.put("contactEmail", businessName.toLowerCase().replace(" ", ".") + "@example.com");
        return application;
    }

    /**
     * Helper method to calculate commission rate based on tier
     */
    private String calculateCommissionRate(String tier) {
        return switch (tier.toUpperCase()) {
            case "BRONZE" -> "6.0%";
            case "SILVER" -> "5.0%";
            case "GOLD" -> "4.0%";
            case "PLATINUM" -> "3.0%";
            default -> "5.0%";
        };
    }

    /**
     * Helper method to get tier benefits
     */
    private List<String> getTierBenefits(String tier) {
        return switch (tier.toUpperCase()) {
            case "BRONZE" -> Arrays.asList("Basic support", "Standard listing");
            case "SILVER" -> Arrays.asList("Priority support", "Featured listing", "Analytics dashboard");
            case "GOLD" -> Arrays.asList("Premium support", "Top listing", "Advanced analytics", "Marketing tools");
            case "PLATINUM" -> Arrays.asList("Dedicated account manager", "Premium placement", "Custom reporting", "API access");
            default -> Arrays.asList("Basic support");
        };
    }
}